pyproject.toml
[tool.poetry]
name = "network-importer"
version = "0.6.0-dev"
description = "Network Importer tool to import an existing network into a Database / Source Of Truth"
authors = ["Damien Garros <dgarros@gmail.com>"]
license = "Apache-2.0"
packages = [
    { include = "network_importer" },
    { include = "dsync" },
]

[tool.poetry.dependencies]
python = "^3.6"
pynetbox = "^4.2"
junos-eznc = "2.2.1"
pybatfish = { git = "git://github.com/batfish/pybatfish.git", branch = "1d1df7e" }
genie = {version = "^19.10", allow-prereleases = true}
jsonschema = {version = "^3.2.0" }
toml = {version = "^0.10.0" }
nornir = "^2.4.0"
termcolor = "^1.1.0"
click = "^7.1.1"
sqlalchemy = "^1.3.17"
pdbpp = "^0.10.2"
pydantic = "^1.6.1"

[tool.poetry.dev-dependencies]
bandit = "^1.6"
pylint = "^2.4"
pyment = "^0.3"
pytest = "^5.4"
requests_mock = "^1.7"
black = {version = "^19.10b0", allow-prereleases = true}

[tool.poetry.scripts]
network-importer = 'network_importer.cli:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
